check for drive formatting for ls command and delete command.

+++ source/scripts/os/shell.ts
@@ -801,12 +801,17 @@ module TSOS {
          */
         public shellDelete(args){
             
+            if(args.length !== 1){
+                _StdOut.putText("Invalid file name");
+                return;
+            }
+            
             var deleted = _krnHDDdriver.deleteFile(args[0]);
             
             if(deleted){
-                _StdOut.putText(args + " has been deleted");
+                _StdOut.putText(args[0] + " has been deleted");
             }else{
-                //TODO:  error checking.
+                _StdOut.putText(args[0] + " has not been deleted.");
             }
             
         }
@@ -839,6 +844,11 @@ module TSOS {
          * lists all the file on the system
          */
         public shellLS(args){
+        
+            if(!_krnHDDdriver.isFormatted){
+                _StdOut.putText("The drive is not formatted, nothing to see here, move along");
+                return;
+            }
             
             if(typeof args[0] !== 'undefined'){
                 _StdOut.putText("Usage: Nothing! Just type ls to list all files");
